refactor: config files
